==> code/5. Scrapping for Streamlit App.ipynb <==
## 5. Scrapping for Streamlit App

The end goal of the project is to build a streamlit app for a user to try out the recommender system.  
For the modeling portion, scikit-surprise only required the username, manga title and user rating.  
Hence, those were the only information collected during the initial scrapping.  
However, to improve user experience on the streamlit app, it is better to display other relevant information about the manga title as well.  
These include a cover image, synopsis and the manga title's specific url to MyAnimeList(MAL) where all information can be found.  
Lastly, the manga title's specific MAL ID number is also scrapped. In case more information is needed, it can be fetched using the Jikan API by calling the specific ID number.

import requests
import pandas as pd
from tqdm import tqdm
from jikanpy import Jikan
import time

jikan = Jikan()

reading_no_zero_df = pd.read_csv('../data/reading_no_zero.csv')

title_list = reading_no_zero_df['item'].unique().tolist()
#title_list = ['Ayashimon','Berserk']
all_info = []
for title in tqdm(title_list):
    try:
        title_info = []
        dict = {}
        dict = jikan.search('manga', title)
        #search title
        title_info.append(title)
        #result title
        title_info.append(dict['results'][0]['title'])
        title_info.append(dict['results'][0]['mal_id'])
        time.sleep(1)
        title_info.append(dict['results'][0]['url'])
        title_info.append(dict['results'][0]['image_url'])
        title_info.append(dict['results'][0]['synopsis'])
        all_info.append(title_info)
        time.sleep(1)
        all_info_df = pd.DataFrame(all_info)
        all_info_df.to_csv('../data/all_info_final.csv', index = False)
    except:
        continue

title_info

all_info_df.rename(columns = {'0':'search title','1':'result title','2':'mal_id','3':'url','4':'image','5':'synopsis'}, inplace = True)
all_info_df.head()

all_info_df.to_csv('../data/all_info_final_2.csv', index = False)
