==> code/6. Scrapping for Streamlit App (continued).ipynb <==
## Scrapping for Streamlit App (continued)

In this notebook, I will be scrapping even more information after some testing of a streamlit app prototype.

### Streamlit App Prototype
I created a prototype app for the recommender system.  
The user searches for and inputs any number of their favorite manga titles via a multiselect box.  
These selected manga titles are then given the maximum rating of 10.  
Recommendations are generated through my machine learning model and displayed to the user.

### Genres
After trying out the streamlit app prototype, I found that even though the genres were displayed in the MyAnimeList(MAL) link, it might be better to display them in the app's results as well.  
By having the genres on hand in my data, I will also be able to allow the user to filter the recommendations based on their preferred genres.  
For example, even if the user inputs multiple Action titles, they will be able to filter the recommendations to show Mystery titles only.  
Lastly, having the genres on hand allows me to set a filter for if the user wishes to see adult recommendations.

### Alternate Titles
Another point of improvement I found from the streamlit app prototype is that I should include alternate titles of the manga titles.  
Manga titles tend to be known by many names other than their English title. E.g. Japanese title, Romanized title etc.  
This will help users to quickly find and input the titles they want.

import requests
import pandas as pd
from tqdm import tqdm
from jikanpy import Jikan
import time

all_info_df = pd.read_csv('../data/all_info_final_2.csv')

all_info_df = all_info_df[['search title','result title','mal_id','url','image','synopsis']]

#likely to contain duplicates as I only have 2.9K titles
all_info_df.shape

all_info_df = all_info_df.drop_duplicates(subset = 'result title')
all_info_df.shape

extra_features_df = all_info_df[['mal_id']]
extra_features_df.shape

#create features in the data frame to store new information
features = ['title_english','title_japanese','title_synonyms', 'genres_1', 'genres_2', 'genres_3', 'genres_4', 'genres_5', 'explicit_genres_1', 'explicit_genres_2', 'explicit_genres_3', 'explicit_genres_4', 'explicit_genres_5', 'demographics_1', 'demographics_2', 'demographics_3', 'themes_1', 'themes_2', 'themes_3']
for feature in features:
    extra_features_df[feature] = 'nil'
extra_features_df.shape

for mal_id in tqdm(extra_features_df['mal_id'].unique().tolist()):
    try:
        res = requests.get(f'https://api.jikan.moe/v4/manga/{mal_id}')
        result = res.json() 
            

        extra_features_df.loc[extra_features_df['mal_id'] == mal_id, ['title_english']] = result['data']['title_english']
        extra_features_df.loc[extra_features_df['mal_id'] == mal_id, ['title_japanese']] = result['data']['title_japanese']
        synonyms = ''
        for name in result['data']['title_synonyms']:
            synonyms = synonyms + '/' + name
            extra_features_df.loc[extra_features_df['mal_id'] == mal_id, ['title_synonyms']] = synonyms

        for i in range(len(result['data']['genres'])):
            extra_features_df.loc[extra_features_df['mal_id'] == mal_id, [f'genres_{i+1}']] = result['data']['genres'][i]['name']


        for i in range(len(result['data']['explicit_genres'])):
            extra_features_df.loc[extra_features_df['mal_id'] == mal_id, [f'explicit_genres_{i+1}']] = result['data']['explicit_genres'][i]['name']

        for i in range(len(result['data']['demographics'])):
            extra_features_df.loc[extra_features_df['mal_id'] == mal_id, [f'demographics_{i+1}']] = result['data']['demographics'][i]['name']

        for i in range(len(result['data']['themes'])):
            extra_features_df.loc[extra_features_df['mal_id'] == mal_id, [f'themes_{i+1}']] = result['data']['themes'][i]['name']
         
        extra_features_df.to_csv('extra_features_df.csv', index = False)
        
    except: 
        continue

extra_features_df = pd.read_csv('../data/extra_features_df.csv')
extra_features_df.head()

#check for and fill null values
extra_features_df['title_english'].isnull().sum()

extra_features_df['title_english'].fillna(value='nil',inplace = True)

extra_features_df['title_english'].isnull().sum()

extra_features_df['title_japanese'].isnull().sum()

extra_features_df['title_japanese'].fillna(value='nil',inplace = True)

extra_features_df['title_japanese'].isnull().sum()

extra_features_df['title_synonyms'].isnull().sum()

extra_features_df.to_csv('../data/extra_features_df_clean.csv', index = False)

extra_features_df = pd.read_csv('../data/extra_features_df_clean.csv')
extra_features_df.head()

all_info_df.columns.unique()

#merge new information with existing information
combined_all_info_df = pd.merge(left = all_info_df, right = extra_features_df, how = 'right', on = 'mal_id')
combined_all_info_df.head()

combined_all_info_df.shape

#check that the 'result title's are exactly the same as what I searched via 'search title'
combined_all_info_df[combined_all_info_df['search title'] != combined_all_info_df['result title']]

#as 'search title's are the same as 'result title's, drop 'search title'
combined_all_info_df.drop(columns = 'search title', inplace = True)
combined_all_info_df.head()

#rename 'result title' to just 'title'
combined_all_info_df.rename(columns = {'result title':'title'}, inplace = True)
combined_all_info_df.head()

combined_all_info_df.columns

def combined_title(row):  
    if row['title_english'] != 'nil':
        if row['title'] != row['title_english']:
            row['title_check'] = str(row['title']) + '/' + str(row['title_english'])
        else:
            row['title_check'] = str(row['title'])
    else:
        row['title_check'] = str(row['title'])
    
    if row['title_japanese'] == 'nil':
        row['title_japanese_check'] = 'nil'
    else:
        row['title_japanese_check'] = str(row['title_japanese'])
    
    if row['title_synonyms'] == 'nil':
        row['title_synonyms_check'] = 'nil'
    else:
        row['title_synonyms_check'] = str(row['title_synonyms'])
    

    if row['title_japanese_check'] == 'nil' and row['title_synonyms_check'] == 'nil':
        return str(row['title_check'])
    elif row['title_japanese_check'] != 'nil' and row['title_synonyms_check'] == 'nil': 
        return str(row['title_check']) + '/' + str(row['title_japanese_check'])
    elif row['title_japanese_check'] == 'nil' and row['title_synonyms_check'] != 'nil':
        return str(row['title_check']) + str(row['title_synonyms_check'])
    else:
        return str(row['title_check']) + '/' + str(row['title_japanese_check']) + str(row['title_synonyms_check'])

combined_all_info_df['combined_title'] = combined_all_info_df.apply(lambda row: combined_title(row), axis=1)
combined_all_info_df.head(10)

combined_all_info_df.to_csv('../data/combined_all_info_df.csv', index = False)

#currently, the genres are stored under columns called 'genres_1', 'genres_2' etc., define a function to combine them
def combined_genres(row):  
    output = ''
    for i in range(1,9):
        output = output + str(row[f'genres_{i}']) + ' '
    return output        

combined_all_info_df['combined_genres'] = combined_all_info_df.apply(lambda row: combined_genres(row), axis=1)
combined_all_info_df.head()

#create a list of all unique genres present in data
all_genres = []
for i in range(1,9):
    for genre in combined_all_info_df[f'genres_{i}'].unique().tolist():
        if genre not in all_genres:
            all_genres.append(genre)
all_genres

#create a new column for each genre
for genre in all_genres:
    combined_all_info_df[genre] = 0
combined_all_info_df.head()

#set genre value to 1 if manga title has that genre, if not, it stays as 0
for ind in combined_all_info_df.index:
     for genre in all_genres:
            if str(genre) in str(combined_all_info_df['combined_genres'][ind]):
                combined_all_info_df[genre][ind] = 1
combined_all_info_df.head()

combined_all_info_df.head()

#filter away irrelevant columns
combined_all_info_df = combined_all_info_df[['title','mal_id','url','image','synopsis','demographics_1', 'demographics_2', 'demographics_3', 'themes_1','themes_2','themes_3','combined_title','Drama','Action', 'Comedy','Sci-Fi','Avant Garde','Horror','Supernatural','Fantasy','Adventure','Romance','Gourmet','Sports','Mystery','Slice of Life','Suspense']]

combined_all_info_df.head()

combined_all_info_df.to_csv('../data/combined_all_info_df_clean.csv', index = False)
